refactor(core): dividir login/page.tsx en secciones (pr2) (#255)
split behavior-preserving de 1420 a 281 lineas. el page queda como
orquestador delgado (datos + mutaciones) y las secciones presentacionales
viven en _components: passwordsection, passwordresetflow, socialconnections,
passkeyssection, activesessionssection. tokeniza colores y tipografia,
mantiene passwordtoggle local. sin cambios de comportamiento.
