+++ frontend/src/app/(tenant)/dashboard/settings/login/_components/index.ts
@@ -14,3 +14,10 @@ export {
   extractErrorMessage,
   maskEmail,
 } from './login-helpers';
+
+export { PasswordSection, type PasswordData } from './PasswordSection';
+export { ChangePasswordForm } from './ChangePasswordForm';
+export { PasswordResetFlow } from './PasswordResetFlow';
+export { SocialConnections } from './SocialConnections';
+export { PasskeysSection } from './PasskeysSection';
+export { ActiveSessionsSection } from './ActiveSessionsSection';
